Patched NodeStore to use extended bits for id space for rels and props

+++ advanced/kernel/src/main/java/org/neo4j/kernel/impl/nioneo/store/PersistenceWindow.java
@@ -32,7 +32,7 @@ public interface PersistenceWindow
      */
     public Buffer getBuffer();
     
-    public Buffer getOffsettedBuffer( int id );
+    public Buffer getOffsettedBuffer( long id );
 
     public int getRecordSize();
